extract images from gallery
